pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "private-copytrading-system"
version = "0.3.0"
description = "TradeNodeX multi-account copy trading control center."
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "aiosqlite>=0.20.0,<1.0.0",
  "boto3>=1.39.0,<2.0.0",
  "cryptography>=45.0.0,<46.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "redis>=6.0.0,<7.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
